corso corsista tuttoooooooo

+++ gestionecorsi/src/com/betacom/businesscomponent/model/CorsoCorsista.java
@@ -1,21 +1,35 @@
 package com.betacom.businesscomponent.model;
 
 public class CorsoCorsista {
-	private int codCorsista;
-	private int codCorso;
-	public int getCodCorsista() {
+	private long codCorsista;
+	private long codCorso;
+	
+
+	
+	public long getCodCorsista() {
 		return codCorsista;
 	}
-	public void setCodCorsista(int codCorsista) {
+
+
+
+	public void setCodCorsista(long codCorsista) {
 		this.codCorsista = codCorsista;
 	}
-	public int getCodCorso() {
+
+
+
+	public long getCodCorso() {
 		return codCorso;
 	}
-	public void setCodCorso(int codCorso) {
+
+
+
+	public void setCodCorso(long codCorso) {
 		this.codCorso = codCorso;
 	}
-	
+
+
+
 	@Override
 	public String toString() {
 		return "CorsoCorsista [codCorsista=" + codCorsista + ", codCorso=" + codCorso + "]";
